Ajout du polymorphisme d'execution pour les différents objets de la scene, il faut encore changer les couleurs qui ne sont pas jolies

+++ Sources/Color.h
@@ -1,4 +1,6 @@
 #pragma once
+#include <cstdint>
+
 
 class Color {
  public:
@@ -16,7 +18,7 @@ class Color {
   Color operator*(double t);
 
  private:
-  int red;
-  int green;
-  int blue;
+  uint8_t red;
+  uint8_t green;
+  uint8_t blue;
 };
